brain-core/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
redis
python-dotenv
google-cloud-firestore
